Further refactor menu UI.
Further moved the creation of the menu away from a pure JS solution, this
time focused on the population of the replay rows in the Replays menu box.
The template row can be edited in html and styled in css more easily and
without having to go into the code.