Start of re-refactor into re-styled modals and a single html file.
The UI html is being reorganized into a single file. The UI styling will
rely on Bootstrap 3.2. To avoid interfering with the style of the TagPro
main page itself, Bootstrap was compiled such that all but the mixins,
variables, print, glyphicons, and modals were nested within a
.bootstrap-container selector. This was selected over a jQuery scoped CSS
plugin as the latter did not function and result in the modals appearing
as expected.

As a result of this scoping, any CSS to style the modals should use the