Refactor replay chat handling. Fixes #104.
Chat messages represent information as it was when the chat occurred,
but we were treating the player/auth/team attribute of them as if they
corresponded to the frame being rendered. To fix this there were 3
changes:

1. When recording a replay, save the name, auth, and team attribute of
   the originating player.
2. Before rendering the replay, pre-process the replay to make the
   format of the chat objects consistent with what we expect. This is
   simpler than handling all cases in the chat rendering itself.
3. When cropping replays, ensure that chat objects are using the new
   format.

Barring a capability to do replay upgrade (as mentioned in #58), this is
the best we can do and should cover all cases.

Also I hid the Renderer behind a function to force consumers of it to
wait on the Promise, which should be easier than remembering to wait on
Renderer#ready.

Testing done:

* Render and preview the replay mentioned in #104.
* Record a new replay and preview/render it, ensuring all expected chats
  were present.
* Preview a few other test replays with known chats.