Add validation code and tests.
Validation takes place in two parts: schema validation and semantic
validation. Schema validation relies on ajv and JSON schema, and
semantic validation is just a function to check the objects.

Currently we have an async validation function since it is easier
to lazy-load the schemas (and requires less coupling between modules),
but if needed in the future we can initialize the schemas ahead of time,
wait on some `init` function of the validate module, and and be completely
synchronous with validation. They may be required for using validation in
Dexie upgrade transactions.