Refactor background page replay handling.
Recently a change was made to bring handling of the details of the replay
data to the background page. This builds on that to consolidate and hide
replay metadata generation/usage from most of the background page.

Fixes #157. We now only grab the replays themselves when existing meta-
data cannot be found. This seems to speed up menu load a good amount but
more testing with larger numbers of replays is needed. For that, #57
should be in place first otherwise it will take a while.

Also fixes #158.