Refactor rendering.
The video rendering function is now isolated from the extension and
independent of the DOM, only depending on blobs to be provided. Whammy
has been edited to take a duration for each frame and the frame source
passed to the video rendering function provides it.

Validation has also been refactored to use named errors instead of
an object with error information.