Move ordering of results to toStream.
Issue #203 revolves around frames being out-of-order. This was due
to the disregard for Response resolution order in WhammyVideo#compile.
Since the expected interface of `map` was to have results returned in
order, it has been replaced with a function that guarantees output in
input order. This also cleans up the rendering a little and increases
test coverage.

Fixes #203.