Add rendering of marsball.
As part of the Easter update we were already tracking entities contained
in `tagpro.objects`, this just adds rendering of objects with type
`marsball`.

The offset is the default for players and objects, including the
marsball. The value is kind of strange so in anticipation of it getting
updated to something else, a `tagproVersion` attribute has also been
added to recorded replays.