Default to a blank tile if map contains an unknown value
Previously, unknown tiles would get mapped to an empty object, and we
would fail later on during replay rendering when trying to use a
property of a property that wasn't there (usually `.coordinates.x`).

Now if there is an unknown tile, we use a blank tile instead. This
should allow replays to continue to be rendered and previewed in the extension
even if new tiles are introduced.