Only report missing floor tiles once during rendering
Originally, a missing floor tile definition would cause an error to be
written for each frame during rendering. This leads to a lot of noise in
the JS console, would be much worse if we tried to proceed with
rendering the current frame, and makes it difficult to centrally send
info about missing tiles.

Now we trace an error once per replay rendered or previewed.